feat: Order nationalities for export and reports
Introduces a new constant `ORDERED_NATIONALITIES_FOR_EXPORT` to define a specific order for nationalities.

This change updates the Reports component to use this new order when generating nationality export files. It also applies the ordered list to the Admin component's reporting to ensure consistent nationality presentation. Nationalities not present in the ordered list will be grouped under "Outros".